📚 docs (docs): updated LICENSE file
This file contains the MIT License for the VolunteerCodeCarajas project. It grants permission to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the software, under certain conditions. The file also includes a third-party license notice for 'wait-for-it.sh' by vishnubob, which is part of the project.